This commit was manufactured by cvs2svn to create tag 'release_26_07_2007_grass5_4_1'.
git-svn-id: https://svn.osgeo.org/grass/grass/tags/release_20070726_grass_5_4_1@9493 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ src.garden/grass.postgresql/d.what.s.pg/runPg.c
@@ -119,7 +119,7 @@ int runPg(SQL_stmt, str_dist, print_out)
     }
 /* help I always screw up atof functions ---  look closely and fix it */
 
-    (double) pts->distance = (double) searchdist;
+    pts->distance = (double) searchdist;
 
     R_open_driver();
     D_setup(0);
